+++ CHANGELOG.md
@@ -2,6 +2,7 @@ You're really going to want to read this.
 
 ## Unreleased
 
+* [*BREAKING*] Cookies now default to secure and http-only.
 * Allow setting custom options (path, domain, secure etc) for an individual cookie.
 * Deprecate Cookie::$salt, ::$secure, etc in favour of a combined Cookie::configure() method
   to set multiple options all at once e.g. from bootstrap.